requirements.txt
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pandas==2.1.3
reportlab==4.0.7
flask==3.0.0
flask-cors==4.0.0
openpyxl==3.1.2

